+++ packages/volto-separator-block/src/components/Separator/View.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const View = ({ data }) => {
-  return <div className="block separator" />;
-};
-
-export default View;