+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking
 
 - Remove Button block in favor of `volto-button-block` @sneridagh
+- Remove Separator block in favor of `volto-separator-block` @sneridagh
 
 ### Feature